Adding support for nested message types.
When building our (sorted) list of module message types, we now visit
any discovered nested types recursively.  We unfortunately need to
perform a small amount of name mangling along the way to conform to
the generated C++ code's naming conventions, but it's not too heinous.

In the module template, the rules for deriving the C++ type name from
a field's type_name attribute have been wrapped up in a Jinja2 macro.